Create README.md
This project mainly aims to create a smart home system that can learn from the user’s daily activities. This smart home system includes an Arduino microcontroller, Ultrasonic sensor, Bluetooth HC-05 chip and an Android phone. The smart home system has a programmed algorithm that takes in input from the Android app and the sensor. This smart home automated system also focused on to reduce human intervention and monitor the smart home system using automatically controlled sensors. The advantage of this system is that it can be expanded and extended by using a sensor and controlling it with Arduino. This system helps in building a more convenient, safe, economic model.